+++ contracts/queue/src/contract.rs
@@ -2,8 +2,8 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 use cosmwasm_std::{
-    from_slice, to_vec, Api, Binary, Env, Extern, Order, Response, Result,
-    Storage,
+    from_slice, to_vec, Api, Binary, Env, Extern, HandleResponse, InitResponse, Order,
+    QueryResponse, Result, Storage,
 };
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -52,15 +52,15 @@ pub fn init<S: Storage, A: Api>(
     _deps: &mut Extern<S, A>,
     _env: Env,
     _msg: InitMsg,
-) -> Result<Response> {
-    Ok(Response::default())
+) -> Result<InitResponse> {
+    Ok(InitResponse::default())
 }
 
 pub fn handle<S: Storage, A: Api>(
     deps: &mut Extern<S, A>,
     env: Env,
     msg: HandleMsg,
-) -> Result<Response> {
+) -> Result<HandleResponse> {
     match msg {
         HandleMsg::Enqueue { value } => enqueue(deps, env, value),
         HandleMsg::Dequeue {} => dequeue(deps, env),
@@ -69,7 +69,11 @@ pub fn handle<S: Storage, A: Api>(
 
 const FIRST_KEY: u8 = 0;
 
-fn enqueue<S: Storage, A: Api>(deps: &mut Extern<S, A>, _env: Env, value: i32) -> Result<Response> {
+fn enqueue<S: Storage, A: Api>(
+    deps: &mut Extern<S, A>,
+    _env: Env,
+    value: i32,
+) -> Result<HandleResponse> {
     // find the last element in the queue and extract key
     let last_key = deps
         .storage
@@ -85,14 +89,14 @@ fn enqueue<S: Storage, A: Api>(deps: &mut Extern<S, A>, _env: Env, value: i32) -
     let data = to_vec(&Item { value })?;
 
     deps.storage.set(&[my_key], &data);
-    Ok(Response::default())
+    Ok(HandleResponse::default())
 }
 
-fn dequeue<S: Storage, A: Api>(deps: &mut Extern<S, A>, _env: Env) -> Result<Response> {
+fn dequeue<S: Storage, A: Api>(deps: &mut Extern<S, A>, _env: Env) -> Result<HandleResponse> {
     // find the first element in the queue and extract value
     let first = deps.storage.range(None, None, Order::Ascending).next();
 
-    let mut res = Response::default();
+    let mut res = HandleResponse::default();
     if let Some((k, v)) = first {
         // remove from storage and return old value
         deps.storage.remove(&k);
@@ -103,26 +107,26 @@ fn dequeue<S: Storage, A: Api>(deps: &mut Extern<S, A>, _env: Env) -> Result<Res
     }
 }
 
-pub fn query<S: Storage, A: Api>(deps: &Extern<S, A>, msg: QueryMsg) -> Result<Vec<u8>> {
+pub fn query<S: Storage, A: Api>(deps: &Extern<S, A>, msg: QueryMsg) -> Result<QueryResponse> {
     match msg {
         QueryMsg::Count {} => query_count(deps),
         QueryMsg::Sum {} => query_sum(deps),
     }
 }
 
-fn query_count<S: Storage, A: Api>(deps: &Extern<S, A>) -> Result<Vec<u8>> {
+fn query_count<S: Storage, A: Api>(deps: &Extern<S, A>) -> Result<QueryResponse> {
     let count = deps.storage.range(None, None, Order::Ascending).count() as u32;
-    to_vec(&CountResponse { count })
+    Ok(Binary(to_vec(&CountResponse { count })?))
 }
 
-fn query_sum<S: Storage, A: Api>(deps: &Extern<S, A>) -> Result<Vec<u8>> {
+fn query_sum<S: Storage, A: Api>(deps: &Extern<S, A>) -> Result<QueryResponse> {
     let values: Result<Vec<Item>> = deps
         .storage
         .range(None, None, Order::Ascending)
         .map(|(_, v)| from_slice(&v))
         .collect();
     let sum = values?.iter().fold(0, |s, v| s + v.value);
-    to_vec(&SumResponse { sum })
+    Ok(Binary(to_vec(&SumResponse { sum })?))
 }
 
 #[cfg(test)]
@@ -142,13 +146,13 @@ mod tests {
 
     fn get_count(deps: &Extern<MockStorage, MockApi>) -> u32 {
         let data = query(deps, QueryMsg::Count {}).unwrap();
-        let res: CountResponse = from_slice(&data).unwrap();
+        let res: CountResponse = from_slice(data.as_slice()).unwrap();
         res.count
     }
 
     fn get_sum(deps: &Extern<MockStorage, MockApi>) -> i32 {
         let data = query(deps, QueryMsg::Sum {}).unwrap();
-        let res: SumResponse = from_slice(&data).unwrap();
+        let res: SumResponse = from_slice(data.as_slice()).unwrap();
         res.sum
     }