+++ packages/std/src/init_handle.rs
@@ -1,0 +1,142 @@
+//! Types and helpers for init and handle
+
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+use crate::encoding::Binary;
+use crate::types::{Coin, HumanAddr};
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "lowercase")]
+pub enum CosmosMsg {
+    // this moves tokens in the underlying sdk
+    Send {
+        from_address: HumanAddr,
+        to_address: HumanAddr,
+        amount: Vec<Coin>,
+    },
+    // this dispatches a call to another contract at a known address (with known ABI)
+    // msg is the json-encoded HandleMsg struct
+    Contract {
+        contract_addr: HumanAddr,
+        msg: Binary, // we pass this in as Vec<u8> to the contract, so allow any binary encoding (later, limit to rawjson?)
+        send: Option<Vec<Coin>>,
+    },
+    // this should never be created here, just passed in from the user and later dispatched
+    Opaque {
+        data: Binary,
+    },
+}
+
+#[derive(Serialize, Deserialize, Clone, Default, Debug, PartialEq, JsonSchema)]
+pub struct LogAttribute {
+    pub key: String,
+    pub value: String,
+}
+
+/// A shorthand to produce log messages
+pub fn log(key: &str, value: &str) -> LogAttribute {
+    LogAttribute {
+        key: key.to_string(),
+        value: value.to_string(),
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Default, Debug, PartialEq, JsonSchema)]
+pub struct InitResponse {
+    // let's make the positive case a struct, it contrains Msg: {...}, but also Data, Log, maybe later Events, etc.
+    pub messages: Vec<CosmosMsg>,
+    pub log: Vec<LogAttribute>, // abci defines this as string
+    pub data: Option<Binary>,   // abci defines this as bytes
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "lowercase")]
+pub enum InitResult {
+    Ok(InitResponse),
+    Err(String),
+}
+
+impl InitResult {
+    // unwrap will panic on err, or give us the real data useful for tests
+    pub fn unwrap(self) -> InitResponse {
+        match self {
+            InitResult::Err(msg) => panic!("Unexpected error: {}", msg),
+            InitResult::Ok(res) => res,
+        }
+    }
+
+    pub fn is_err(&self) -> bool {
+        match self {
+            InitResult::Err(_) => true,
+            _ => false,
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Default, Debug, PartialEq, JsonSchema)]
+pub struct HandleResponse {
+    // let's make the positive case a struct, it contrains Msg: {...}, but also Data, Log, maybe later Events, etc.
+    pub messages: Vec<CosmosMsg>,
+    pub log: Vec<LogAttribute>, // abci defines this as string
+    pub data: Option<Binary>,   // abci defines this as bytes
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "lowercase")]
+pub enum HandleResult {
+    Ok(HandleResponse),
+    Err(String),
+}
+
+impl HandleResult {
+    // unwrap will panic on err, or give us the real data useful for tests
+    pub fn unwrap(self) -> HandleResponse {
+        match self {
+            HandleResult::Err(msg) => panic!("Unexpected error: {}", msg),
+            HandleResult::Ok(res) => res,
+        }
+    }
+
+    pub fn is_err(&self) -> bool {
+        match self {
+            HandleResult::Err(_) => true,
+            _ => false,
+        }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use crate::{coin, from_slice, to_vec};
+
+    #[test]
+    fn can_deser_error_result() {
+        let fail = InitResult::Err("foobar".to_string());
+        let bin = to_vec(&fail).expect("encode contract result");
+        println!("error: {}", std::str::from_utf8(&bin).unwrap());
+        let back: InitResult = from_slice(&bin).expect("decode contract result");
+        assert_eq!(fail, back);
+    }
+
+    #[test]
+    fn can_deser_ok_result() {
+        let send = InitResult::Ok(InitResponse {
+            messages: vec![CosmosMsg::Send {
+                from_address: HumanAddr("me".to_string()),
+                to_address: HumanAddr("you".to_string()),
+                amount: coin("1015", "earth"),
+            }],
+            log: vec![LogAttribute {
+                key: "action".to_string(),
+                value: "release".to_string(),
+            }],
+            data: None,
+        });
+        let bin = to_vec(&send).expect("encode contract result");
+        println!("ok: {}", std::str::from_utf8(&bin).unwrap());
+        let back: InitResult = from_slice(&bin).expect("decode contract result");
+        assert_eq!(send, back);
+    }
+}