+++ packages/std/src/lib.rs
@@ -2,6 +2,8 @@
 
 mod encoding;
 mod errors;
+mod init_handle;
+mod query;
 mod serde;
 mod storage;
 mod traits;
@@ -13,15 +15,17 @@ pub use crate::errors::{
     contract_err, dyn_contract_err, invalid, unauthorized, ApiError, Error, NotFound, NullPointer,
     ParseErr, Result, SerializeErr,
 };
+pub use crate::init_handle::{
+    log, CosmosMsg, HandleResponse, HandleResult, InitResponse, InitResult, LogAttribute,
+};
+pub use crate::query::{QueryResponse, QueryResult};
 pub use crate::serde::{from_slice, to_vec};
 pub use crate::storage::MemoryStorage;
 pub use crate::traits::{Api, Extern, ReadonlyStorage, Storage};
 #[cfg(feature = "iterator")]
 pub use crate::traits::{Order, KV};
 pub use crate::transactions::{transactional, transactional_deps, RepLog, StorageTransaction};
-pub use crate::types::{
-    coin, log, CanonicalAddr, ContractResult, CosmosMsg, Env, HumanAddr, QueryResult, Response,
-};
+pub use crate::types::{coin, CanonicalAddr, Env, HumanAddr};
 
 // Exposed in wasm build only