+++ packages/vm/src/testing.rs
@@ -6,7 +6,7 @@ use serde::Serialize;
 use schemars::JsonSchema;
 
 use cosmwasm_std::testing::{mock_dependencies, MockApi, MockStorage};
-use cosmwasm_std::{to_vec, Api, ContractResult, Env, QueryResult, Storage};
+use cosmwasm_std::{to_vec, Api, Env, HandleResult, InitResult, QueryResult, Storage};
 
 use crate::calls::{call_handle, call_init, call_query};
 use crate::compatability::check_wasm;
@@ -32,9 +32,9 @@ pub fn init<S: Storage + 'static, A: Api + 'static, T: Serialize + JsonSchema>(
     instance: &mut Instance<S, A>,
     env: Env,
     msg: T,
-) -> ContractResult {
+) -> InitResult {
     match to_vec(&msg) {
-        Err(e) => ContractResult::Err(e.to_string()),
+        Err(e) => InitResult::Err(e.to_string()),
         Ok(serialized_msg) => call_init(instance, &env, &serialized_msg).unwrap(),
     }
 }
@@ -46,9 +46,9 @@ pub fn handle<S: Storage + 'static, A: Api + 'static, T: Serialize + JsonSchema>
     instance: &mut Instance<S, A>,
     env: Env,
     msg: T,
-) -> ContractResult {
+) -> HandleResult {
     match to_vec(&msg) {
-        Err(e) => ContractResult::Err(e.to_string()),
+        Err(e) => HandleResult::Err(e.to_string()),
         Ok(serialized_msg) => call_handle(instance, &env, &serialized_msg).unwrap(),
     }
 }